requirements.txt
langchain==0.0.149
pdf2image

openai
pinecone-client
unstructured
pdfminer.six
python-dotenv
tiktoken
flask
flask_restful
psycopg2

DateTime==5.1
newsapi-python==0.2.7
numpy==1.24.2
pandas==2.0.0
pandas_ta==0.2.45b
textblob==0.17.1
ipython==8.12.0

hugchat
youtube-transcript-api
schedule
google-api-python-client
pytesseract